pyproject.toml
[tool.poetry]
name = "gifts-api"
version = "0.1.0"
description = ""
authors = ["Arthur Moreira de Deus <arthurdedeus@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
asgiref = "3.7.2"
build = "1.0.3"
CacheControl = "0.13.1"
certifi = "2024.2.2"
cffi = "1.16.0"
charset-normalizer = "3.3.2"
cleo = "2.1.0"
crashtest = "0.4.1"
distlib = "0.3.8"
Django = "5.0.2"
djangorestframework = "3.14.0"
fastjsonschema = "2.19.1"
filelock = "3.13.1"
idna = "3.6"
importlib-metadata = "7.0.1"
installer = "0.7.0"
keyring = "24.3.1"
msgpack = "1.0.8"
packaging = "23.2"
pexpect = "4.9.0"
pkginfo = "1.9.6"
platformdirs = "3.11.0"
poetry = "1.7.1"
poetry-core = "1.8.1"
poetry-plugin-export = "1.6.0"
ptyprocess = "0.7.0"
pycparser = "2.21"
pyproject_hooks = "1.0.0"
pytz = "2024.1"
rapidfuzz = "3.6.1"
requests = "2.31.0"
requests-toolbelt = "1.0.0"
shellingham = "1.5.4"
sqlparse = "0.4.4"
urllib3 = "2.0.7"
virtualenv = "20.25.1"
python = "^3.11"
psycopg2 = "^2.9.9"
factory-boy = "^3.3.0"
pillow = "^10.2.0"
gunicorn = "^21.2.0"
whitenoise = "^6.6.0"
django-environ = "^0.11.2"
django-storages = "^1.14.2"
boto3 = "^1.34.51"
django-cors-headers = "^4.3.1"
pypix = {git = "https://github.com/cleitonleonel/pypix.git"}

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.5.0"
ruff = "0.2.2"

[tool.ruff]
exclude = [
    ".ruff_cache",
    ".venv",
    "__pypackages__",
    "venv",
]

line-length = 88
indent-width = 4

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I", "DJ", "S", "C90"]
ignore = ["DJ001"]

fixable = ["ALL"]
unfixable = []

dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
